first pass at reorganizing code
