+++ test/diagrammatic-programs/diagrams.jl
@@ -1,0 +1,3 @@
+using Test
+
+
